+++ e2e/cjs-config/__tests__/test.js
@@ -5,10 +5,6 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {Config} from '@jest/types';
-
-export type ProjectPackageJson = {
-  jest?: Partial<Config.InitialOptions>;
-  scripts?: Record<string, string>;
-  type?: 'commonjs' | 'module';
-};
+test('dummy test', () => {
+  expect(1).toBe(1);
+});