+++ e2e/cjs-config/jest.config.cjs
@@ -5,10 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {Config} from '@jest/types';
-
-export type ProjectPackageJson = {
-  jest?: Partial<Config.InitialOptions>;
-  scripts?: Record<string, string>;
-  type?: 'commonjs' | 'module';
+module.exports = {
+  displayName: 'Config from cjs file',
+  testEnvironment: 'node',
 };