+++ packages/jest-cli/src/init/__tests__/fixtures/has_jest_config_file_cjs/jest.config.cjs
@@ -5,10 +5,4 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {Config} from '@jest/types';
-
-export type ProjectPackageJson = {
-  jest?: Partial<Config.InitialOptions>;
-  scripts?: Record<string, string>;
-  type?: 'commonjs' | 'module';
-};
+module.exports = {};