+++ packages/jest-cli/src/init/constants.ts
@@ -6,4 +6,12 @@
  */
 
 export const PACKAGE_JSON = 'package.json';
-export const JEST_CONFIG = 'jest.config.js';
+export const JEST_CONFIG_BASE_NAME = 'jest.config';
+export const JEST_CONFIG_EXT_CJS = '.cjs';
+export const JEST_CONFIG_EXT_JS = '.js';
+export const JEST_CONFIG_EXT_JSON = '.json';
+export const JEST_CONFIG_EXT_ORDER = Object.freeze([
+  JEST_CONFIG_EXT_JS,
+  JEST_CONFIG_EXT_CJS,
+  JEST_CONFIG_EXT_JSON,
+]);